requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2
django-bootstrap-v5==1.0.11
django-bootstrap5==23.1
django-distill==3.1.3
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
filelock==3.12.0
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
platformdirs==3.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.29.0
soupsieve==2.4.1
sqlparse==0.4.4
urllib3==1.26.15
virtualenv==20.23.0
